Add release scripts
tools/release/release.sh will now ask for version number and
generates a commit for it. The version number was moved to a
specific bzl file to be easier to manipulate.

tools/release/release.yaml is a Google Cloud Container Builder
script that will build the update site and publish it to eclipse.bazel.build/updatesite

This script only tag a release, it does not mark it as a release
on GitHub yet.

Change-Id: I4711f2552d876bd0a74cd24d6408da85f629a905